Changed dynamic programming problem to Fibonacci.
There doesn't seem to be a good dynamic programming problem which
requires a two dimensional table, but doesn't take a significant amount
of thought to solve.

A dynamic programming (well, just memoization) approach to finding the
n'th number in the Fibonacci sequence should be a much easier problem to
digest while completing an exam.