Run pdflatex twice for each rule, to evaluate references.
It's not necessary always, but if we're building from scratch or using
new references, we need to run pdflatex twice.

If we want to only do this extra work when necessary, we could check to
see if %.aux (which stores concrete data for the references) changes
after running `pdflatex' the first time.