Fixed stuff in NullPointerException problem
a) I prefer the term "object" over `Object` as the latter ties it into Java's `Object` type, which `null` isn't specifically related to.
b) Example given doesn't throw a NullPointerException (NPE); it is an error at compile time since it is inside a function. I changed `myList` to be an instance variable (where NPE's do occur). I included this in the answer itself, as I feel it is important to note.
c) "instantiate" means to create an object, which is *needed* to avoid NPE's, not the opposite.
d) Diamond operator was introduced in Java7. The reason I removed "type inference" was because I didn't want to imply that Java7 had fantastic type inference or anything (even though this is type inference).